Get trips working

+++ src/main/java/com/uom/trips/tripsrepository/TripsRepository.java
@@ -3,6 +3,6 @@ package com.uom.trips.tripsrepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.uom.trips.com.uom.trips.model.Trip;
 
-public interface TripsRepository extends JpaRepository<Trip, Long>{
+public interface TripsRepository extends JpaRepository<Trip, Integer>{
 
 }
